requirements.txt
pandas
numpy
pyyaml
requests
pyarrow
statsmodels
scikit-learn
xgboost
matplotlib
